do not use radio buttons for 'single' choice filter
when a variable has 3 levels or less, gfilter uses radio buttons when 
using 'single' choice selector. Although this is generally a good idea, 
it becomes less useful when the levels are very long. Unfortunately in 
this case the widget eats too much horizontal space. Using a combobox 
all the time solves the issue.